refactor axes selector
